@@ -1,6 +1,6 @@
 {
   "name": "mkdocs-material",
-  "version": "5.1.1",
+  "version": "5.1.6",
   "description": "A Material Design theme for MkDocs",
   "keywords": [
     "mkdocs",
@@ -41,6 +41,7 @@
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^5.13.0",
+    "@mdi/svg": "^5.1.45",
     "@primer/octicons": "^9.6.0",
     "@types/clipboard": "^2.0.1",
     "@types/copy-webpack-plugin": "^5.0.0",
@@ -49,46 +50,46 @@
     "@types/html-minifier": "^3.5.3",
     "@types/lunr": "^2.3.3",
     "@types/mini-css-extract-plugin": "^0.9.1",
-    "@types/node": "^13.13.0",
-    "@types/ramda": "^0.27.3",
+    "@types/node": "^13.13.4",
+    "@types/ramda": "^0.27.4",
     "@types/resize-observer-browser": "^0.1.3",
     "@types/webpack": "^4.41.12",
     "@types/webpack-assets-manifest": "^3.0.0",
     "autoprefixer": "^9.7.6",
     "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^3.5.2",
-    "css-mqpacker": "^7.0.0",
+    "css-loader": "^3.5.3",
     "event-hooks-webpack-plugin": "^2.1.6",
     "expose-loader": "^0.7.5",
     "github-types": "^1.0.0",
     "gitlab": "^14.2.2",
     "html-minifier": "^4.0.0",
     "ignore-emit-webpack-plugin": "^2.0.2",
+    "imagemin-webpack-plugin": "^2.4.2",
     "material-design-color": "^2.3.2",
-    "material-design-icons-svg": "^3.0.0",
     "material-shadows": "^3.0.1",
     "mini-css-extract-plugin": "^0.9.0",
     "modularscale-sass": "^3.0.10",
+    "postcss-inline-svg": "^4.1.0",
     "postcss-loader": "^3.0.0",
-    "preact": "^10.4.0",
+    "postcss-svgo": "^4.0.2",
+    "preact": "^10.4.1",
     "resize-observer-polyfill": "^1.5.1",
     "rimraf": "^3.0.2",
-    "sass": "^1.26.3",
+    "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
-    "string-replace-loader": "^2.3.0",
-    "stylelint": "^13.3.2",
+    "stylelint": "^13.3.3",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-order": "^4.0.0",
-    "stylelint-scss": "^3.17.0",
-    "ts-loader": "^7.0.1",
-    "ts-node": "^8.8.2",
+    "stylelint-scss": "^3.17.1",
+    "ts-loader": "^7.0.2",
+    "ts-node": "^8.10.1",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
     "tslib": "^1.11.1",
     "tslint": "^5.20.1",
     "tslint-sonarts": "^1.9.0",
     "typescript": "^3.8.3",
-    "webpack": "^4.42.1",
+    "webpack": "^4.43.0",
     "webpack-assets-manifest": "^3.1.1",
     "webpack-cli": "^3.3.11"
   },
