@@ -26,7 +26,9 @@ import {
   distinctUntilChanged,
   filter,
   map,
+  mapTo,
   pluck,
+  startWith,
   switchMap
 } from "rxjs/operators"
 
@@ -34,6 +36,7 @@ import { WorkerHandler, watchElementOffset } from "browser"
 import {
   SearchMessage,
   SearchResult,
+  isSearchReadyMessage,
   isSearchResultMessage
 } from "integrations"
 
@@ -70,6 +73,13 @@ export function mountSearchResult(
     switchMap(el => {
       const container = el.parentElement!
 
+      /* Compute if search is ready */
+      const ready$ = rx$
+        .pipe(
+          filter(isSearchReadyMessage),
+          mapTo(true)
+        )
+
       /* Compute whether there are more search results to fetch */
       const fetch$ = watchElementOffset(container)
         .pipe(
@@ -85,7 +95,8 @@ export function mountSearchResult(
         .pipe(
           filter(isSearchResultMessage),
           pluck("data"),
-          applySearchResult(el, { query$, fetch$ })
+          applySearchResult(el, { query$, ready$, fetch$ }),
+          startWith([])
         )
     })
   )
