@@ -24,13 +24,16 @@
 // which must be tackled after we gathered some feedback on v5.
 // tslint:disable
 
-import { values } from "ramda"
+import { sortBy, prop, values } from "ramda"
 import {
   merge,
   combineLatest,
   animationFrameScheduler,
   fromEvent,
-  from
+  from,
+  defer,
+  of,
+  NEVER
 } from "rxjs"
 import { ajax } from "rxjs/ajax"
 import {
@@ -42,7 +45,9 @@ import {
   observeOn,
   take,
   shareReplay,
-  pluck
+  pluck,
+  catchError,
+  map
 } from "rxjs/operators"
 
 import {
@@ -228,60 +233,71 @@ export function initialize(config: unknown) {
 
   /* ----------------------------------------------------------------------- */
 
-  // External index
-  const index = config.search && config.search.index
-    ? config.search.index
-    : undefined
+  /* Search worker */
+  const worker$ = defer(() => {
+    const index = config.search && config.search.index
+      ? config.search.index
+      : undefined
 
-  /* Fetch index if it wasn't passed explicitly */
-  const index$ = typeof index !== "undefined"
-    ? from(index)
-    : base$
-        .pipe(
-          switchMap(base => ajax({
-            url: `${base}/search/search_index.json`,
-            responseType: "json",
-            withCredentials: true
-          })
-            .pipe<SearchIndex>(
-              pluck("response")
+    /* Fetch index if it wasn't passed explicitly */
+    const index$ = typeof index !== "undefined"
+      ? from(index)
+      : base$
+          .pipe(
+            switchMap(base => ajax({
+              url: `${base}/search/search_index.json`,
+              responseType: "json",
+              withCredentials: true
+            })
+              .pipe<SearchIndex>(
+                pluck("response")
+              )
             )
           )
-        )
 
-  const worker = setupSearchWorker(config.search.worker, {
-    base$, index$
+    return of(setupSearchWorker(config.search.worker, {
+      base$, index$
+    }))
   })
 
   /* ----------------------------------------------------------------------- */
 
   /* Mount search query */
-  const query$ = useComponent("search-query")
+  const search$ = worker$
     .pipe(
-      mountSearchQuery(worker, { transform: config.search.transform }),
-      shareReplay(1)
-    )
+      switchMap(worker => {
 
-  /* Mount search reset */
-  const reset$ = useComponent("search-reset")
-    .pipe(
-      mountSearchReset(),
-      shareReplay(1)
-    )
+        const query$ = useComponent("search-query")
+          .pipe(
+            mountSearchQuery(worker, { transform: config.search.transform }),
+            shareReplay(1)
+          )
 
-  /* Mount search result */
-  const result$ = useComponent("search-result")
-    .pipe(
-      mountSearchResult(worker, { query$ }),
-      shareReplay(1)
-    )
+        /* Mount search reset */
+        const reset$ = useComponent("search-reset")
+          .pipe(
+            mountSearchReset(),
+            shareReplay(1)
+          )
 
-  /* ----------------------------------------------------------------------- */
+        /* Mount search result */
+        const result$ = useComponent("search-result")
+          .pipe(
+            mountSearchResult(worker, { query$ }),
+            shareReplay(1)
+          )
 
-  const search$ = useComponent("search")
-    .pipe(
-      mountSearch({ query$, reset$, result$ }),
-      shareReplay(1)
+        return useComponent("search")
+          .pipe(
+            mountSearch(worker, { query$, reset$, result$ }),
+            shareReplay(1)
+          )
+      }),
+      catchError(() => {
+        useComponent("search")
+          .subscribe(el => el.hidden = true) // TODO: Hack
+        return NEVER
+      })
     )
 
   /* ----------------------------------------------------------------------- */
@@ -337,8 +353,50 @@ export function initialize(config: unknown) {
       })
 
   /* Enable instant loading, if not on file:// protocol */
-  if (config.features.includes("instant") && location.protocol !== "file:")
-    setupInstantLoading({ document$, location$, viewport$ })
+  if (config.features.includes("instant") && location.protocol !== "file:") {
+
+    /* Fetch sitemap and extract URL whitelist */
+    base$
+      .pipe(
+        switchMap(base => ajax({
+          url: `${base}/sitemap.xml`,
+          responseType: "document",
+          withCredentials: true
+        })
+          .pipe<Document>(
+            pluck("response")
+          )
+        ),
+        withLatestFrom(base$),
+        map(([document, base]) => {
+          const urls = getElements("loc", document)
+            .map(node => node.textContent!)
+
+          // Hack: This is a temporary fix to normalize instant loading lookup
+          // on localhost and Netlify previews. If this approach proves to be
+          // suitable, we'll refactor URL whitelisting anyway. We take the two
+          // shortest URLs and determine the common prefix to isolate the
+          // domain. If there're no two domains, we just leave it as-is, as
+          // there isn't anything to be loaded anway.
+          if (urls.length > 1) {
+            const [a, b] = sortBy(prop("length"), urls)
+
+            /* Determine common prefix */
+            let index = 0
+            while (a.charAt(index) === b.charAt(index))
+              index++
+
+            /* Replace common prefix (i.e. base) with effective base */
+            for (let i = 0; i < urls.length; i++)
+              urls[i] = urls[i].replace(a.slice(0, index), `${base}/`)
+          }
+          return urls
+        })
+      )
+        .subscribe(urls => {
+          setupInstantLoading(urls, { document$, location$, viewport$ })
+        })
+  }
 
   /* ----------------------------------------------------------------------- */
 
